requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.8
distro==1.9.0
dnspython==2.4.2
email_validator==2.2.0
exceptiongroup==1.1.3
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Markdown==3.8
MarkupSafe==2.1.1
mongoengine==0.29.1
openai==1.78.0
packaging==23.1
pillow==11.1.0
pluggy==1.3.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.0.2
python-dotenv==0.19.2
requests==2.27.1
sentinels==1.0.0
sniffio==1.3.1
tomli==2.0.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.9
waitress==2.1.2
Werkzeug==2.3.7
WTForms==3.0.1
